+++ lib/services/local/embeddings.go
@@ -18,6 +18,7 @@ import (
 	"context"
 	"time"
 
+	aiembeddings "github.com/gravitational/teleport/lib/ai/embeddings"
 	"github.com/gravitational/trace"
 	"github.com/jonboulle/clockwork"
 	"github.com/sirupsen/logrus"
@@ -26,7 +27,6 @@ import (
 	"github.com/gravitational/teleport/api/utils/retryutils"
 	"github.com/gravitational/teleport/lib/ai"
 	"github.com/gravitational/teleport/lib/backend"
-	"github.com/gravitational/teleport/lib/services"
 )
 
 // EmbeddingsService implements the services.Embeddings interface.
@@ -48,7 +48,7 @@ func (e EmbeddingsService) GetEmbedding(ctx context.Context, kind, resourceID st
 	if err != nil {
 		return nil, trace.Wrap(err)
 	}
-	return services.UnmarshalEmbedding(result.Value)
+	return aiembeddings.UnmarshalEmbedding(result.Value)
 }
 
 // GetEmbeddings returns a stream of embeddings for a given kind.
@@ -56,7 +56,7 @@ func (e EmbeddingsService) GetEmbeddings(ctx context.Context, kind string) strea
 	startKey := backend.ExactKey(embeddingsPrefix, kind)
 	items := backend.StreamRange(ctx, e, startKey, backend.RangeEnd(startKey), 50)
 	return stream.FilterMap(items, func(item backend.Item) (*ai.Embedding, bool) {
-		embedding, err := services.UnmarshalEmbedding(item.Value)
+		embedding, err := aiembeddings.UnmarshalEmbedding(item.Value)
 		if err != nil {
 			e.log.Warnf("Skipping embedding at %s, failed to unmarshal: %v", item.Key, err)
 			return nil, false
@@ -67,7 +67,7 @@ func (e EmbeddingsService) GetEmbeddings(ctx context.Context, kind string) strea
 
 // UpsertEmbedding creates or update a single ai.Embedding in the backend.
 func (e EmbeddingsService) UpsertEmbedding(ctx context.Context, embedding *ai.Embedding) (*ai.Embedding, error) {
-	value, err := services.MarshalEmbedding(embedding)
+	value, err := aiembeddings.MarshalEmbedding(embedding)
 	if err != nil {
 		return nil, trace.Wrap(err)
 	}
